+++ src/Disturbance_PnET.cs
@@ -37,9 +37,9 @@ namespace Landis.Extension.Succession.PnETForC
             BiomassReductionParameters.ToList().ForEach(x => parameters.Add("disturbance:"+x.Key, x.Value));
         }
 
-        public static float ReduceFoliage(float FolBiomass, double defoliationFrac)
+        public static double ReduceFoliage(double FolBiomass, double defoliationFrac)
         {
-            FolBiomass *= (float)(1.0 - defoliationFrac);
+            FolBiomass *= (double)(1.0 - defoliationFrac);
             return FolBiomass;
         }
 
@@ -47,15 +47,15 @@ namespace Landis.Extension.Succession.PnETForC
         {
             if (sitecohorts == null)
                 throw new Exception("sitecohorts should not be null");
-            float WoodDebrisReductionFrac = 0;
-            float LeafLitterReductionFrac = 0;
+            double WoodDebrisReductionFrac = 0;
+            double LeafLitterReductionFrac = 0;
             if (disturbanceType != null && Names.TryGetParameter(disturbanceType.Name, out Parameter<string> parameter))
             {
                 // If parameters are available, then set the loss fractions here.
                 if (parameter.ContainsKey("WoodDebrisReduction"))
-                    WoodDebrisReductionFrac = float.Parse(parameter["WoodDebrisReduction"]);
+                    WoodDebrisReductionFrac = double.Parse(parameter["WoodDebrisReduction"]);
                 if (parameter.ContainsKey("LeafLitterReduction"))
-                    LeafLitterReductionFrac = float.Parse(parameter["LeafLitterReduction"]);
+                    LeafLitterReductionFrac = double.Parse(parameter["LeafLitterReduction"]);
             }
             ((SiteCohorts)sitecohorts).RemoveWoodDebris(WoodDebrisReductionFrac);
             ((SiteCohorts)sitecohorts).RemoveLeafLitter(LeafLitterReductionFrac);
@@ -66,23 +66,23 @@ namespace Landis.Extension.Succession.PnETForC
             if (sitecohorts == null)
                 throw new Exception("sitecohorts should not be null");
             // By default, all material is allocated to the wood debris or the leaf litter pool
-            float WoodReductionFrac = 0;
-            float RootReductionFrac = 0;
-            float FolReductionFrac = 0;
+            double WoodReductionFrac = 0;
+            double RootReductionFrac = 0;
+            double FolReductionFrac = 0;
             if (disturbanceType != null && Names.TryGetParameter(disturbanceType.Name, out Parameter<string> parameter))
             {
                 // If parameters are available, then set the loss fractions here.
                 if (parameter.ContainsKey("WoodReduction"))
-                    WoodReductionFrac = float.Parse(parameter["WoodReduction"]);
+                    WoodReductionFrac = double.Parse(parameter["WoodReduction"]);
                 if (parameter.ContainsKey("RootReduction"))
-                    RootReductionFrac = float.Parse(parameter["RootReduction"]);
+                    RootReductionFrac = double.Parse(parameter["RootReduction"]);
                 if (parameter.ContainsKey("FolReduction"))
-                    FolReductionFrac = float.Parse(parameter["FolReduction"]);
+                    FolReductionFrac = double.Parse(parameter["FolReduction"]);
             }            
             // Add new dead wood and leaf litter
-            float woodAdded = (float)((1 - WoodReductionFrac) * cohort.Wood * frac);
-            float rootAdded = (float)((1 - RootReductionFrac) * cohort.Root * frac);
-            float folAdded = (float)((1 - FolReductionFrac) * cohort.Fol * frac);
+            double woodAdded = (double)((1 - WoodReductionFrac) * cohort.Wood * frac);
+            double rootAdded = (double)((1 - RootReductionFrac) * cohort.Root * frac);
+            double folAdded = (double)((1 - FolReductionFrac) * cohort.Fol * frac);
             // Using Canopy fractioning
             ((SiteCohorts)sitecohorts).AddWoodDebris(woodAdded * cohort.CanopyLayerFrac, cohort.PnETSpecies.WoodDebrisDecompRate);
             ((SiteCohorts)sitecohorts).AddWoodDebris(rootAdded * cohort.CanopyLayerFrac, cohort.PnETSpecies.WoodDebrisDecompRate);
